Criação do Form

+++ SalePharma/src/main/java/br/edu/iff/ccc/bsi/webdev/repository/ProdutoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.iff.ccc.bsi.webdev.modelo.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 
 }
